+++ opencog/atoms/reduct/AssignLink.cc
