+++ opencog/atoms/reduct/AssignLink.h
